Written a code
This code will choose a random word from the wordlist and print underscore so that word will be hidden. and also give idea about the word length. taking input from the user/player later checking whether the guessed letter is in  choosenword using for loop and inside it a if-else condition.